Use GPIOnn macros when setting peripheral pins
* changed: periperhal API which sets pins is now using GPIOxy macros
  instead bare 0 .. 31 values. This makes possible to use single place
  to set where peripheral is mapped.

+++ include/libopencm3/nrf/common/i2c.h
@@ -122,7 +122,7 @@ void i2c_set_standard_mode(uint32_t i2c);
 void i2c_set_frequency(uint32_t i2c, uint32_t freq);
 void i2c_send_data(uint32_t i2c, uint8_t data);
 uint8_t i2c_get_data(uint32_t i2c);
-void i2c_select_pins(uint32_t i2c, uint8_t scl_pin, uint8_t sda_pin);
+void i2c_select_pins(uint32_t i2c, uint32_t scl_pin, uint32_t sda_pin);
 void i2c_set_address(uint32_t i2c, uint8_t addr);
 void i2c_resume(uint32_t i2c);
